+++ src/main/java/com/ds/watchtable/entity/Pos.java
@@ -24,9 +24,9 @@ public class Pos extends BaseEntity{
     private String drink1;
     private String drink2;
 
-    private int table2;
-    private int table4;
-    private int table6;
+    private Long table2;
+    private Long table4;
+    private Long table6;
 
     @ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
     private Store store;
